+++ foxtive/src/database/mod.rs
@@ -2,9 +2,14 @@ use diesel::r2d2::ConnectionManager;
 use diesel::{r2d2, PgConnection};
 use serde::Serialize;
 
+mod config;
+mod conn;
 pub mod ext;
 mod ext_impl;
 pub mod pagination;
+
+pub use config::DbConfig;
+pub use conn::create_db_pool;
 
 pub type DBPool = r2d2::Pool<ConnectionManager<PgConnection>>;